+++ run_developer_setup.sh
@@ -4,4 +4,5 @@ pip install -r requirements.txt
 
 # Install netharn in developer mode
 #pip install -e .
-python setup.py clean && python setup.py develop
+#python setup.py clean && python setup.py develop
+pip install -e .